requirements.txt
beautifulsoup4==4.11.1
boto3==1.24.27
botocore==1.27.27
certifi==2022.6.15
charset-normalizer==2.1.0
idna==3.3
jmespath==1.0.1
kaggle==1.5.12
numpy==1.23.1
pandas==1.4.3
pickle-mixin==1.0.2
python-dateutil==2.8.2
python-dotenv==0.20.0
python-slugify==6.1.2
pytz==2022.1
requests==2.28.1
s3transfer==0.6.0
six==1.16.0
smart-open==6.0.0
soupsieve==2.3.2.post1
text-unidecode==1.3
tqdm==4.64.0
urllib3==1.26.10
zipfile36==0.1.3
